Create the initial game class.
The commit implements an intial game class capable of creating an SDL window on the screen.

The class implements the basic game loop and enables a means of exiting the application via keypress.

The code in this commit is likely largely the same as the code from the initial lessons in the Pikuma 2D C++ course.