Merge mousemap and tilemap types.
The change moves the functionality of the mousemap type into the tilemap. The mousemap as a standalone type is gone.

The code at this commit will break in the case that the cursor moves way outside (I think to the left of?) the playable screen space. Likely because of division by zero or negative number handling.

Also, the code does not correctly get the map grid from a specified screen pixel.