feat: update camera for display mode
Purpose of the changes in this commit is to update the camera such that the width and height of the camera are now based on the display_mode (SDL_DisplayMode) and not constant height and width values.

This has involved making the Game::camera member a shared pointer which is created during the initialise() method.