feat: bounding box
This commit introduces changes necessary to render a bounding box around vertical sprites in debug mode.

Additionally, it separates the concept of the mouse map from the mouse such that the mousemap can be used to convert a pixel coordinate to a grid coordinate separately from the Mouse instance.

The next thing to investigate is creating a means to easily access aspects of a sprite's position in the world.

The challenge here is that the world position for an entity with a sprite component is determined by:

1. The entity's transform position; modified by
2. Sprite offsets, stored in the sprite component

A calculation is performed each frame to determine the destination SDL_Rect to which a sprite should be rendered - should this be a component in itself?