Streamline render logic
The changes in this commit reduce the total number of loops involved in rendering sprites by combining the render_bounding_box and render_sprite functions such that sprites are only looped once, with bounding boxes optionally rendered based on a boolean function argument.

The code for sprite rendering has been consolidated into a single, templated function which is functionally identical for the different sprite types (Terrain, Vertical).

Realistically, it should be possible to have a single sprite component type with an additional attribute (Z index), and to do away with the need to cater for sprites of types.

Additionally, the changes on this branch do not yet comprehensively address the map out-of-bounds issue because it's not currently easily possible to determine the 'bottom middle' point of a moving sprite.

Further changes should be introduced as the subject of separate tickets to address these issues.