Some changes to delta time
This commit introduces some changes to the Game::update() function to make the purpose of the function a little clearer.

In effect, the function should:

1. Figure out when the previous frame started; calculate the difference between then and now as a delta time
2. Do some updates
3. Check how long the updates took
4. Wait for a while, in the case that the updates took less time than a a frame should

There are also some changes introduced which allow the use of binary operators (multiplication) of GLM vector types, which should make things (e.g. use of velocity) more straightforward.