Document GLM vector usage and update vector usage in IMGUI
Purpose of this change is to document the instances when GLM vectors are used, the types used (e.g. vec2 vs ivec2), and rationale for the use of the given type.

This was done because it's possible that incorrect use of GLM types could cause stuttering (though I don't think that's what's going on).

Additionally I've changed the types (ivec2 -> vec2) used to create position and velocity components for new entities created via IMGUI.