Streamline spatial constants
The purpose of this commit is to reduce the total number of constants relating to points in space by replacing individual x, y-based constants with glm vectors.

For example, the TILE_WIDTH and TILE_HEIGHT constants have been replaced by a single TILE_SIZE constant which has an x and y component.

In the short term this will reduce the total number of constants and likely make the existing code easier to read.

In the medium term, code which utilises these constants can (should) be updated to make better use of vector operations (addition, multiplication, subtraction, etc).