feat: spritesheets
This commit enables the creation of a new spritesheet type which encapsulates:
 - The spritesheet texture
 - The SDL_Rect(s) specifying the invidual sprite positions within the sheet

Additionally, the 'old' ('home-made') assets have been replaced with free-to-use alternatives by Kenney.

Finally, some code referencing the loading of the 'old' sprite textures has been replaced (or, for now, commented out).

This commit breaks the ability of the game engine to show images on the screen; some work is left to:

 - Update the sprites so that they can refer to the spritesheet texture they 'belong' to
 - Update the game logic to load sprites from the spritesheet using the SDL_Rect sprite positions