Use std::optional members appropriately
The main point of this commit is to change an unordered map of spritesheets into two individual spritesheet variables.

There was basically no point for the time being in using an unordered map for the spritesheets, because:
 - There are only two spritesheets; and so
 - Accessing individual spritesheets (building tiles, city tiles) via a map seemed a bit excessing

*excessive.

Using varibles in place of an unordered map threw up some issues is using std::optional.

Specifically, it became apparent that 'late' initialising optional values used copy constructors (at least in the case of the SpriteSheets).

Because the spritesheets had pointer members, the copy constructor resulted in the creation of a new SpriteSheet whose 'spritesheet' member pointed to memory which had already been freed.

To mitigate this, the std::optional method  has been used to construct optional variables in place - mitigating the need for the copy.

This was a good lesson!

The next step should be to check the 'const correctness' of the SpriteSheet type - in particular whether SDL_Texture* and SDL_Rect& members can / should be const and whether the member functions used to return them can also be const.