Update Camera:: to return ivec2
The Camera::get_position() member was returning a SDL_Rect&, which includes a position AND a height/width; the height/width were never relevant for instances where the get_position() was being called, and so an ivec2 would do instead.

This has left a few gaps where ivec2 operations (+, -) could be implemented a little more simply; to be updated in further commits!