feat: prototype for bitwise connection indicators
The prototype proves some logic which could be used to indicate the connecting edges for a single tile.

The logic is based on using the right-most 4 bits of an integer to indicate whether the tile has a road connecting any of the North, East, South or West edges.

Each time an update is made to a tile where it is indicated that the tile has connecting edges, some logic can be used to check whether the tile adjacent to each connecting edge reciprocates the connection.

For example, if a tile is updated to indicate that it has a connection to the WEST (left), the left-adjacent tile is checked to see whether it has a corresponding connection to the EAST (right).

Corresponding connections result in a match, which can be the trigger for some other logic - e.g. creating or updating nodes/edges in the graph.