feat: some code to enable edge intersection
This commit introduces some code necessary to enable edge intersection using the bitwise and operator.

Specifically, it introduces a reference to the 'parent
' tilemap in the graph & associated edges.

This is necessary because edge intersection needs to retrieve a tile from an index position in the graph in order to create new edges.

New edges are created between the origin and termination points of intersecting edges and the intersection point.

In essence, when two edges cross over one another, FOUR edges are created:
 - edge one to crossing point
 - crossing point to edge one end
 - edge two to crossing point
 - crossing point to edge two end